src/langprop/requirements.txt
numpy
matplotlib
torch
torchvision
torchaudio
tqdm
wandb
einops
typing_extensions
python-dotenv
langchain[llms]
timeout-decorator
